Remove InputAccessoryView causing crash on android
