+++ src/Navigator.tsx
@@ -14,8 +14,8 @@ import { DrawerContent } from './components/DraweContent/DrawerContent';
 import { ComposeScreen } from './screens/ComposeScreen';
 import { TestScreen } from './screens/TestScreen';
 import { colors } from './util/colors';
-import { RegisterConsentScreen } from './screens/RegisterConsentScreen';
-import { RegisterUsernameScreen } from './screens/RegisterUsernameScreen';
+import { RegisterConsentScreen } from './screens/RegisterConsent/RegisterConsentScreen';
+import { RegisterUsernameScreen } from './screens/RegisterUsername/RegisterUsernameScreen';
 import { RegisterPasswordScreen } from './screens/RegisterPasswordScreen';
 import { RegisterRecoveryEmailScreen } from './screens/RegisterRecoveryEmailScreen';
 import { RegisterSuccessScreen } from './screens/RegisterSuccessScreen';
