+++ src/util/platform.ts
@@ -1,0 +1,4 @@
+import { Platform } from 'react-native';
+
+export const isIOS = Platform.OS === 'ios';
+export const isAndroid = Platform.OS === 'android';
